chore: move crd to classes
