+++ .htaccess
@@ -52,7 +52,7 @@
     RewriteRule ^kirby/(.*) index.php [L]
 
     # block direct access to starter classes
-    RewriteRule ^src/(.*) index.php [L]
+    RewriteRule ^classes/(.*) index.php [L]
 
     # block all files in the assets folder from being accessed directly
     RewriteRule ^assets/(.*) index.php [L]
