+++ composer.json
@@ -24,7 +24,7 @@
   },
   "autoload": {
     "psr-4": {
-      "KirbyStarter\\": "src/"
+      "KirbyStarter\\": "classes/"
     }
   },
   "config": {
