+++ package.json
@@ -20,10 +20,10 @@
   "devDependencies": {
     "autoprefixer": "^10.4.21",
     "fluid-tailwind": "^1.0.4",
-    "glob": "^11.0.1",
+    "glob": "^11.0.2",
     "postcss": "^8.5.3",
     "tailwindcss": "^3.4.17",
-    "vite": "^6.2.4",
+    "vite": "^6.3.3",
     "vite-plugin-kirby": "^5.4.0"
   },
   "scripts": {
